requirements.txt
python-telegram-bot==20.7
langchain>=0.3
langchain-openai>=0.2
python-dotenv